+++ crates/matrix-sdk-base/src/notification_settings.rs
@@ -19,10 +19,15 @@ use serde::{Deserialize, Serialize};
 /// Enum representing the push notification modes for a room.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
 pub enum RoomNotificationMode {
-    /// Receive notifications for all messages.
+    /// Receive remote and in-app notifications for all messages.
     AllMessages,
-    /// Receive notifications for mentions and keywords only.
-    MentionsAndKeywordsOnly,
+    /// Receive remote and in-app notifications for mentions and keywords only.
+    MentionsAndKeywordsOnly {
+        /// If true, receive in-app-only notifications (no pushes) for other
+        /// room messages. Otherwise, mute all other room messages.
+        #[cfg(feature = "unstable-msc3768")]
+        notify_in_app: bool,
+    },
     /// Do not receive any notifications.
     Mute,
 }