+++ crates/matrix-sdk-base/CHANGELOG.md
@@ -22,6 +22,9 @@ All notable changes to this project will be documented in this file.
   `RoomInfo::invite_details` method returns both the timestamp and the
   inviter.
   ([#5390](https://github.com/matrix-org/matrix-rust-sdk/pull/5390))
+- Add a new field `notify_in_app` in `RoomNotificationMode::MentionsAndKeywordsOnly`
+  behind a new feature `unstable-msc3768`.
+  ([#5441](https://github.com/matrix-org/matrix-rust-sdk/pull/5441
 
 ### Refactor
 - [**breaking**] The `Stripped` variants of `RawAnySyncOrStrippedTimelineEvent`,